Modularize endpoints and add transactions for writing endpoints.
